rename lib to nameify
