+++ nameify/nameifier.go
@@ -1,4 +1,4 @@
-package lib
+package nameify
 
 import (
 	_ "embed"
