+++ nameify/nameifier_test.go
@@ -1,4 +1,4 @@
-package lib
+package nameify
 
 import (
 	"testing"
